requirements.txt
torrentool
requests
bitarray
PyYAML
